Display popup when two people reserve same spot
