+++ index.html
@@ -2,7 +2,7 @@
 <html lang="fr" translate="no">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/png" sizes="16x16" href="/public/presences.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="/presences.png">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Présences</title>
   </head>
